+++ server/router/my.ts
@@ -4,8 +4,9 @@ import jwtMiddleware from '../middleware/jwtMiddleware';
 import wrapAsync from '../utils/wrapAsync';
 
 const myRouter: Router = express.Router();
-import { getMyReview } from './../controllers/review';
+import { getMyReview, getMyQuestion } from './../controllers/my';
 
 myRouter.get('/review', jwtMiddleware, wrapAsync(getMyReview));
+myRouter.get('/question', jwtMiddleware, wrapAsync(getMyQuestion));
 
 export default myRouter;
